+++ eggs/naaya.content.meeting/naaya/content/meeting/participants.py
@@ -89,8 +89,8 @@ class Participants(SimpleItem):
     def _set_attendee(self, uid, role):
         def can_set_role():
             participants_count = self.participantsCount()
-            if (meeting.max_participants > participants_count
-                    or meeting.max_participants == 0):
+            if (meeting.max_participants > participants_count or
+                    meeting.max_participants == 0):
                 return True
             # can also change rights even if meeting is full
             if meeting.max_participants == participants_count:
@@ -344,7 +344,7 @@ class Participants(SimpleItem):
         survey = self.get_survey()
         if survey:
             for answer in survey.objectValues('Naaya Survey Answer'):
-                if answer.respondent in [uid, 'signup:'+uid]:
+                if answer.respondent in [uid, 'signup:' + uid]:
                     return True
         return False